requirements.txt
requests==2.31.0
pytest==8.0.0
jsonschema==4.21.1
curlify==2.2.1
allure-pytest==2.13.2
selene==2.0.0rc4
python-dotenv==1.0.1